// rokim_todo_ui/src/app/services/todo/model/todo.ts
/**
 * rokim_todo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Todo { 
    category: string;
    created_at: string;
    description?: string | null;
    done_at?: string | null;
    id: number;
    title: string;
}